+++ main.tf
@@ -10,12 +10,13 @@ locals {
   create        = var.use_existing_slb ? false : var.create
   create_others = var.use_existing_slb || var.create ? true : false
   this_slb_id   = var.use_existing_slb ? var.existing_slb_id : var.create ? concat(alicloud_slb.this.*.id, [""])[0] : ""
+  address_type  = var.address_type != "" ? var.address_type : var.internal == false ? "internet" : "intranet"
 }
 resource "alicloud_slb" "this" {
   count                = local.create ? 1 : 0
   name                 = var.name
   internet_charge_type = var.internet_charge_type
-  address_type         = var.internal == false ? "internet" : "intranet"
+  address_type         = local.address_type
   vswitch_id           = var.vswitch_id
   specification        = var.spec == "" ? null : var.spec
   bandwidth            = var.bandwidth